kern: Add a simple slab allocator implementation
xv6 has only a page allocator and cannot dynamically allocate objects smaller
than a page. To work around this limitation, xv6 declares objects smaller than
a page statically. For example, xv6 declares an array of file structures, an
array of proc structures, and so on. As a result, the number of files the
system can have open is limited by the size of the statically declared file
array, which has NFILE entries (see kernel/file.c and kernel/param.h).

The solution is to adopt the slab allocator.

The slab allocator builds on the page allocator and manages objects of a
specific size (e.g., file structures). It maintains a number of "slabs". For
simplicity, we assume that a slab spans exactly one page. Each slab contains
some metadata at the beginning and a number of objects after the metadata.

For allocation, the slab allocator finds a slab that is not full and allocates
a free object from the slab. If all slabs are full, the slab allocator may ask
the page allocator for more free slabs (via kalloc). For deallocation, the slab
allocator gives an object back to the corresponding slab. If all objects within
a slab are free, the slab allocator may give the slab back to the page allocator
(via kfree).

This implementation attempts to mimic the slab allocator discussed in Jeff
Bonwick's "The Slab Allocator: An Object-Caching Kernel Memory Allocator". This
paper can be found at:
https://www.usenix.org/legacy/publications/library/proceedings/bos94/bonwick.html

Signed-off-by: Tyler Fanelli <tfnll@protonmail.com>