kern: Implement copy-on-write fork()
The goal of copy-on-write (COW) fork() is to defer allocating and copying
physical memory pages for the child until the copies are actually needed, if
ever.

COW fork() creates just a pagetable for the child, with PTEs for user memory
pointing to the parent’s physical pages. COW fork() marks all the user PTEs in
both parent and child as read-only. When either process tries to write one of
these COW pages, the CPU will force a page fault. The kernel page-fault handler
detects this case, allocates a page of physical memory for the faulting process,
copies the original page into the new page, and modifies the relevant PTE in the
faulting process to refer to the new page, this time with the PTE marked
writeable. When the page fault handler returns, the user process will be able to
write its copy of the page.

Signed-off-by: Tyler Fanelli <tfnll@protonmail.com>