kern/alarm: Add sigalarm() and sigreturn() syscalls
This feature periodically alerts a process as it uses CPU time. This might be
useful for compute-bound processes that want to limit how much CPU time they
use, or for processes that want to compute but also want to take some periodic
action.

A system call, sigalarm(interval, handler) is introduced. If an application
calls sigalarm(n, fn), then after every n “ticks” of CPU time that the program
consumes, the kernel should cause application function fn to be called. When fn
returns, the application should resume where it left off. A tick is a fairly
arbitrary unit of time in xv6, determined by how often a hardware timer
generates interrupts.

Another system call, sigreturn(), is also introduced. User alarm handlers are
required to call the sigreturn system call when they have finished. sigreturn's
main responsibility is to restore the process's register context to that of
which the process was executing before the sigalarm handler was executed.

Signed-off-by: Tyler Fanelli <tfnll@protonmail.com>