kern/kalloc: Implement per-CPU freelists
The root cause of lock contention in kalloctest is that kalloc() has a single
free list, protected by a single lock. To remove lock contention, this commit
redesigns the memory allocator to avoid a single lock and list. The basic idea
is to maintain a free list per CPU, each list with its own lock. Allocations and
frees on different CPUs can run in parallel, because each CPU will operate on a
different list. In the case in which one CPU’s free list is empty, but another
CPU’s list has free memory, the one CPU must “steal” part of the other CPU’s
free list. Stealing may introduce lock contention, but that will hopefully be
infrequent.

Signed-off-by: Tyler Fanelli <tfnll@protonmail.com>